비행기록과 영상의 시간 차 조절시 float 단위로 조절 되게끔 수정

+++ center/js/center.js
@@ -1380,7 +1380,7 @@ function flightDetailInit(target) {
 		            return;
 		        }
 		
-		        curVal = parseInt(curVal);        		        
+		        curVal = parseFloat(curVal);        		        
 		        updateFlightRecordDsec(target, curVal);
 		        
 		        showAlert(GET_STRING_CONTENT("msg_sync_adjusted") + " : " + curVal + GET_STRING_CONTENT("label_second"));
